fix: prevent integer overflow in density calculations for large graphs
CommunityDetector.getDensity() and SubgraphExtractor.getDensity() computed
n*(n-1) using int arithmetic, which overflows for communities/subgraphs with
more than ~46,341 vertices. Cast n to long before multiplication to match
the pattern already used in GraphUtils and GraphGenerator.

+++ Gvisual/src/gvisual/CommunityDetector.java
@@ -80,7 +80,7 @@ public class CommunityDetector {
         public double getDensity() {
             int n = members.size();
             if (n < 2) return 0.0;
-            return (2.0 * internalEdges) / (n * (n - 1));
+            return (2.0 * internalEdges) / ((long) n * (n - 1));
         }
 
         /**
